run/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
google-cloud-vision==2.6.3
google-cloud-storage==1.43.0
firebase-admin
